+++ src/main.js
@@ -1,9 +1,3 @@
-import { ScoreDriver } from '@/metrics/score_driver'
-import { Frontend } from '@/ui/frontend.js'
-
-function main() {
-    const score_driver = new ScoreDriver()
-    const frontend = new Frontend(score_driver, document)
-}
-
-main()
+/* v8 ignore next 2 */
+import { run } from './run.js'
+run()